add feature_reduction.py

+++ src/download_images_from_mongo.py
@@ -11,6 +11,7 @@ client = pymongo.MongoClient()
 db = client['flickr_links']
 s3 = boto3.resource('s3')
 
+IMAGE_DIR = '../images/'
 
 def download_image(link):
     """
@@ -19,15 +20,16 @@ def download_image(link):
 
     Opens the link and downloads the file to disk. Uploads to s3
     """
-    print link, bucket_name
-    return
 
     response = urllib2.urlopen(link)
-    with open(link[36:], 'wb') as outfile:
+    with open(IMAGE_DIR + link[36:], 'wb') as outfile:
         outfile.write(response.read())
-    s3.Object(bucket_name, link[36:]).put(Body=open(link[36:]))
-    os.remove(link[36:])
-
+    if bucket_name is None:
+        return
+    else:
+        s3.Object(bucket_name, link[36:]).put(Body=open(link[36:]))
+        os.remove(link[36:])
+    return
 
 def url_generator(collection):
     '''
@@ -42,12 +44,17 @@ if __name__ == '__main__':
         coll_name = sys.argv[1]
         global bucket_name
         bucket_name = sys.argv[2]
+        s3.create_bucket(Bucket=bucket_name)
+    elif len(sys.argv) == 2:
+        coll_name = sys.argv[1]
+        global bucket_name
+        bucket_name = None
     else:
         print """Please provide a string for the MongoDB collection and a string for the s3 bucket"""
         sys.exit(1)
 
-    s3.create_bucket(Bucket=bucket_name)
     collection = db[coll_name]
 
-    pool = Pool(2)
+    url_list = url_generator(collection)
+    pool = Pool()
     results = pool.map(download_image, url_list)
